+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 
+- Spatial index support for useCartoLayerProps [#425](https://github.com/CartoDB/carto-react/pull/425)
 - Ensure source exists in HistogramWidget before getting stats [#426](https://github.com/CartoDB/carto-react/pull/426)
 
 ## 1.3